+++ Assets/_GameAssets/Scripts/Helpers/ActionTypes.cs
@@ -1,0 +1,7 @@
+public static class ActionTypes
+{
+    public const string MOVE = "IsMoving";
+    public const string SLIDE = "IsSliding";
+    public const string SLIDE_START = "IsSlidingActive";
+    public const string JUMP = "IsJumping";
+}