+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.0.7] - 2025-05-23
+
+### Added
+- Added `ActionTypes` class to define player action types.
+- Added `PlayerAnimationController` for managing and updating the current player animations.
+- Added `GetCurrentSpeed()` function to `PlayerController` class for accessing current speed.
+- Added `SetMovementSpeed()` function to `PlayerController` class for handling player's current speed.
+
 ## [0.0.6] - 2025-05-23
 
 ### Added
@@ -10,6 +18,7 @@
 - Updated `PlayerController` for state handling logic to use new `PlayerState` enum values.
 
 ## [0.0.5] - 2025-05-23
+
 ### Added
 [+] Added camera follow system for 3rd person character.
 
@@ -17,6 +26,7 @@
 [~] Fixed character rotation problem.
 
 ## [0.0.4] - 2025-05-23
+
 ### Added
 [+] Added project starter files.
 [+] Added character move/slide.
@@ -24,13 +34,16 @@
 [+] Added character movement speed limit.
 
 ## [0.0.3] - 2025-05-22
+
 ### Added
 [+] Completed the project setup.
 
 ## [0.0.2] - 2025-05-22
+
 ### Added
 [+] Required Unity packages have been added.
 
 ## [0.0.1] - 2025-05-21
+
 ### Added
 [+] Required Unity project files have been added.